requirements/common.txt
django==1.8.2
jsonfield==1.0.3
requests==2.2.1
django-cors-headers==1.1.0
djangorestframework==3.1.2
djangorestframework-jwt==1.6.0
django-haystack==2.4.0
django-allauth==0.21.0
django-filter==0.11.0
PyJWT==1.4.0
django-braces
django-autofixture
django-ckeditor==5.0.2
django-spaghetti-and-meatballs==0.1.1
django-debug-toolbar==1.4
coverage==4.0.3
pytest-django==2.9.1
django-import-export==0.4.2
flake8
XlsxWriter==0.9.3
`